Eliminate all GCC warnings

+++ src/Scope.cpp
@@ -65,7 +65,7 @@ void Scope::plot(
     std::vector<float>& plotY,
     std::vector<float>& plotNormal)
 {
-    for (int i = 0; i < plotY.size(); i++) {
+    for (int i = 0; i < static_cast<int>(plotY.size()); i++) {
         float thicknessX = std::sin(plotNormal[i]) * m_thicknessInPixels / g_windowWidth * 0.5;
         float thicknessY = std::cos(plotNormal[i]) * m_thicknessInPixels / g_windowHeight * 0.5;
         m_coordinates[4 * i + 0] = 2 * plotX[i] - 1 - thicknessX;
@@ -79,7 +79,7 @@ void Scope::plotFilled(
     std::vector<float>& plotX,
     std::vector<float>& plotY)
 {
-    for (int i = 0; i < plotY.size(); i++) {
+    for (int i = 0; i < static_cast<int>(plotY.size()); i++) {
         m_coordinates[4 * i + 0] = 2 * plotX[i] - 1;
         m_coordinates[4 * i + 1] = 2 * plotY[i] - 1;
         m_coordinates[4 * i + 2] = 2 * plotX[i] - 1;
